--- src/Equipe.java
@@ -1,33 +1,0 @@
-public class Equipe {
-
-	private String codinome;
-
-	private int quantidade;
-
-	private double latitude;
-
-	private double longitude;
-
-	public Equipe(String codinome, int quantidade, double latitude, double longitude) {
-		this.codinome = codinome;
-		this.quantidade = quantidade;
-		this.latitude = latitude;
-		this.longitude = longitude;
-	}
-
-	public String getCodinome() {
-		return codinome;
-	}
-
-	public int getQuantidade() {
-		return quantidade;
-	}
-
-	public double getLatitude() {
-		return latitude;
-	}
-
-	public double getLongitude() {
-		return longitude;
-	}
-}
