+++ src/Main.java
@@ -1,3 +1,5 @@
+import telas.Programa;
+
 public class Main {
     public static void main(String[] args) {
         new Programa();
